reorg project, do first 3 templates
