+++ main.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 	"strings"
 
-	"./src/sinit"
+	"./sinit"
 )
 
 func main() {
@@ -61,14 +61,15 @@ func main() {
 	fmt.Println("Name: " + projName)
 	fmt.Println("Stack: " + stack)
 
-	absPath, metaData := sinit.CreateProject(projName)
+	// absPath, metaData := sinit.CreateProject(projName)
+	sinit.CreateProject(projName)
 
-	switch stack {
-	case "node":
-		sinit.InitNode(absPath, metaData)
-	case "hugo":
-		sinit.InitHugo(absPath, theme, themeRepo, deploy, metaData)
-	}
+	// switch stack {
+	// case "node":
+	// 	sinit.InitNode(absPath, metaData)
+	// case "hugo":
+	// 	sinit.InitHugo(absPath, theme, themeRepo, deploy, metaData)
+	// }
 
 	fmt.Printf("\nProject Created. Just type `cd %v`", projName)
 }
