+++ sinit/utils.go
@@ -49,17 +49,7 @@ func runInteractiveCmdFromDir(dir string, cmd string, args ...string) error {
 	return command.Run()
 }
 
-func createFileFromTemplate(templateFile string, ouputFilename string, data interface{}) error {
-	templateDir, err := filepath.Abs("templates")
-	if err != nil {
-		return err
-	}
-
-	tmpl, err := template.New(templateFile).ParseFiles(path.Join(templateDir, templateFile))
-	if err != nil {
-		return err
-	}
-
+func createFileFromTemplate(tmpl *template.Template, ouputFilename string, data interface{}) error {
 	newFile, err := os.Create(ouputFilename)
 	if err != nil {
 		return err
