--- templates/codeowners
@@ -1,2 +1,0 @@
-# All files should be reviewed by {{.Author}}
-* {{.Email}}
