--- templates/readme.md
@@ -1,2 +1,0 @@
-# {{.Name}}
-The `{{.Name}}` project is authored by {{.Author}}
